Use new Yahoo Finance URL
Yahoo changed their URL scheme to require an established session with
cookies. The 'curl' package handles all the details. First establish a
connection to the main site to get a cookie, then use that to get a
crumb, and use the crumb in the download URL. Note that times also
changed to POSIX/UNIX timestamps.